requirements.txt
pandas
sqlalchemy
seaborn
fastapi
python-dotenv
unittest
httpx